Allow buildkit-built images to execute (#103)
Depending on the Docker version, some images may have labels in either
the Config or the ContainerConfig field.  The ContainerConfig field may also
have valid keys but empty values or be missing entirely depending on the version.

This patch takes the labels from both fields (with Config overriding ContainerConfig if
there is a clash).

https://docs.docker.com/engine/deprecated/#container-and-containerconfig-fields-in-image-inspect

Co-authored-by: Jeremy Tuloup <jeremy.tuloup@gmail.com>